server/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "toml (==0.10.2)",
    "loguru (>=0.6.0,<0.7.0)",
    "starlette (>=0.45.3,<0.46.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "aioboto3 (>=13.4.0,<14.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "cryptography (>=44.0.0,<45.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
